Make the display of the warning "Useless use of a constant (%s)" more robust
If the constant is a string (POK), we dump it using pv_pretty, to
properly escape non-printable characters. This also improves detection
of utf-8 encoding in the constant to be printed. Also, this patch
streamlines the tests for the type and value of that constant.

Potential backwards-compatibility issues : if the constant is a string
it will now be enclosed in double quotes, like this :
    Useless use of a constant ("a") in void context
instead of
    Useless use of a constant (a) in void context