mktables: Preserve old format in some tables
Future commits will cause tables that map to code points to, in general,
use deltas instead.  This ensures that files that contain tables and
have been mentioned publicly in the past continue to have their current
contents and format, so that applications that read them (such as
Unicode::Normalize) are unaffected.