Restore directory extension and version in fileify under EFS.
In df2786654 and 8a5aa89570, the traditional behavior of adding
the .DIR;1 onto a fileified directory spec was removed when
operating under Extended Filename Syntax.  Various scary comments
were added about its being a bug to add a type and version onto
a Unix-style path, but actually the CRTL appears to be perfectly
happy with, for example:

   stat('/foo/bar/baz.dir;1');

and without the extension, the home-grown rmdir() fails in the
case of a directory with no preceding path information.  E.g.,

   rmdir('foo');

was failing because there was no internal translation to foo.dir
before passing it to SYS$ERASE.

Moreover, even if there were something wrong with adding .DIR;1,
it has nothing to do with EFS.