mktables: Change generated file comment
All the files that mktables generates that are for external-to-core use
have now been changed so that the code requests explicitly for each that
they have the comment that says they are for external use, but it is
deprecated to use them.  That means that any files that haven't been so
explicitly set should have the comment instead that says they are for
internal use only.