Move typemap documentation to its own file
Sadly, the POD in Typemap.xs was not easily extractable into a POD file
at build time, so it now lives in a separate POD file from the start.
Makes keeping documentation and testing efforts in sync marginally
harder, but it's probably the right trade-off.

What's left to do is finding the right places in other POD files to
refer to this old/new documentation.