mktables: Generate some delta tables
This commit has the effect of changing the non-legacy tables for the lc,
uc, tc, and fc properties to use maps of deltas from the code points
instead of the code points themselves, thus shortening them
significantly, and hence the time required to search through them.

Note that these tables are new, and currently used only by Unicode::UCD.
A future commit will change the Perl core to use them.