bisect.pl now treats the command line as files to match with --match.
Previously with --match it would ignore the command line options, and search
in all distribution files (no --target given), or the build products only
(--target given). These remain the defaults if there are no command line
arguments. But now if command line arguments are given, this are used as glob
patterns to give the list of files to match against.