bisect.pl now exits with 0 for success for --help
bisect.pl --help now exits with 0 for success. The "that took" timing
message is no longer shown after the help text. The sanity test for "are we
running from the current checkout" is now moved after the argument checking,
so that bisect.pl --help will always show the help text.