Remove dead and obscure typemap entries
T_DATAUNIT and T_CALLBACK are nowhere to be found in a CPAN module and
are not used in core. Their purpose is entirely unclear and they are
trivial. They'll always be available from CPAN from the
ExtUtils::Typemaps::Excommunicated module. See perlxstypemap for details
on how to use that if you need it.