Add compile-time inversion lists for POSIX classes
These will be used in regcomp.c to replace the existing bit-wise
handling of these, enabling subsequent optimizations.

These are compiled-in, and hence affect the memory footprint of every
program, including those that don't use Unicode.  The lists that aren't
tiny are therefore currently restricted to only the Latin1 range;
anything needed beyond that will have to be read in at execution time,
just as before.

The design allows for easy conversion from Latin1 to use the full
Unicode range, should it be deemed desirable for some or all of these.