Deprecate is_utf8_char()
This function assumes that there is enough space in the buffer to read
however many bytes are indicated by the first byte in the alleged UTF-8
encoded string.  This may not be true, and so it can read beyond the
buffer end.  is_utf8_char_buf() should be used instead.