In perlfunc, standardise the descriptions for keywords enabled by features.
Previously the "Functions by Category" summaries noted that state, __SUB__
and the "switch" keywords are enabled only by the feature system, but made
no mention of this for evalbytes, fc and say. __SUB__ didn't note that it
was also always available with a CORE:: prefix. Now the descriptions are
consistent for all feature-enabled keywords.