Skip stat.t readability test on VMS.
This new test grants full access on a test file to group and other.
Since I'm a member of my own group and everyone is someone, I of
course have full access to the file, but the test expects it to be
unreadable.  If there is a hierarchy in Unix file permissions, it
clearly doesn't work the same way as the VMS one.