bisect.pl now reports a meaningful error for certain "can't start" scenarios.
If --end is specified but not --start, then don't probe for start revisions
that are more recent than the --end. Previously bisect.pl would test all
stable revisions for a start point, and if it happened to find a (seemingly)
valid start point, it would continue on to run git bisect, which would fail
with the error:

    Some good revs are not ancestor of the bad rev.
    git bisect cannot work properly in this case.
    Maybe you mistake good and bad revs?

which doesn't make clear what the problem actually is. Now the error is:

    Can't find a suitable start revision to default to.
    Tried perl-5.002 perl-5.003 perl-5.004 perl-5.005 perl-5.6.0 at ...