Remove the never-working mktemp() fallback code for FAKE_BIT_BUCKET
A working /dev/null is needed for the implementation of -e
Catamount doesn't have a /dev/null, so it sets FAKE_BIT_BUCKET to enable
emulation code, which fakes the bit bucket with a very short lived temporary
file.

FAKE_BIT_BUCKET had conditionally complied code to create the temporary file
with mkstemp() if the platform has it, with the *intent* to fall back to
mktemp() otherwise. However, the mktemp() code did not work (and could never
have done so), but the problems were concealed by a combination of things.
The code was never actually compiled in, as it was guarded by
#ifdef HAS_MKTEMP, a CPP symbol that the core never defines. The core has
assumed since perl-1.0 that mktemp() is always available. However, because
the code (if run) only changes scriptname from BIT_BUCKET, it wasn't easily
noticed that it didn't run, because on any platform with a valid working
BIT_BUCKET, the subsequent open (of it, not the tempfile) would succeed, and
nothing would fail. When the CPP conditionals are removed, such that the
code runs, it then *fails*, because mktemp() only creates a file *name*, but
doesn't actually create it on disk, so the subsequent open fails.