Add t/run/fdopen.t, which tests the /dev/fd/\d+ code for opening scripts.
perl uses S_open_script() in perl.c to open the file handle on the Perl
program. In 5.003 this was enhanced to parse filenames such as /dev/fd/3
(directly using fdopen() to open file descriptor 3), as part of a suidperl
fix. It was subsequently further enhanced to permit $0 to be set by
suffixing the /dev/fd/\d+ with the file name to use.

Although suidperl was removed from the core by commit cc69b689ee7c2745 in
Jan 2009, this code remained, as it's a useful feature for anyone wishing
to maintain suidperl externally, or write an alternative secure wrapper
system.