In perl.c, only compile S_forbid_setid() if it's needed.
If SETUID_SCRIPTS_ARE_SECURE_NOW is defined, S_forbid_setid() is a no-op,
so don't compile it. Move the declaration and definition of S_forbid_setid()
into the same pre-processor blocks as are used for S_validate_suid(), which
is a no-op when SETUID_SCRIPTS_ARE_SECURE_NOW is /not/ defined.