load Algorithm::Diff later in corelist-perldelta
corelist-perldelta.pl is a bit schizophrenic. The part to generate the
new Perldelta text does not need Algorithm::Diff, but wants to be
run with the freshly built Perl. Ideally, the two usages get
clarified and split up into two separate programs. The added documentation
tries to give two helpful usages of the program.