charnames.t: Attempt to fix probable timing issue
charnames.t sometimes fails on a test that is expecting a file to be
non-existent.  However, other tests in it briefly create a file by that
name.  When two instances of the .t are run in parallel, as happens in
some smokes, it could be that the file exists at the time it is expected
not to.  The result is a non-reproducible failure.  Anyway, that is my
theory.  And so this commit changes to use a filename that is used just
by the test for what happens when a file is non-existent, and hence
doesn't ever exist even transiently.