mktables: Don't duplicate entries
The Name table can have multiple names for each code point.  The highest
priority ones are first in the file.  Prior to this patch, adding a
high priority name to a code point which already had the same name, the
old name could be retained, leaving two identical names for the code
point.  This patch causes the lower-priority one to be deleted.