Copy the pv in parse_label
Two commits ago, the label-handling code in the tokenizer was changed
to record the label in an SVOP instead of just a plain unclaimed PV.
Likewise, the parser (perly.y) was changed to copy the string out of
the SVOP’s SV, insteading of taking ownership of the existing PV.
parse_label was modified to take the string out of the SVOP, but was
not copying it like perly.y.  So it resulted in this:

$ PERL_DESTRUCT_LEVEL=2 ./perl -Ilib ext/XS-APItest/t/swaplabel.t
1..56
ok 1
... truncated ...
ok 56
panic: free from wrong pool during global destruction.